entropy.db: add support for Portage 2.2_rc29 feature "DEFINED_PHASES"
This is a new metadatum which is going to be stored in Portage's vartree
representing the list of defined ebuild phases for the package. For us
it comes into help to choose if to pull up Portage on package install or
not, resulting in a huge speed improvement.

This commit only adds database support, more to come.