entropy.client.interfaces.methods: complete support for in-memory repositories.
In-memory repositories support was partially implemented some time
ago. This commit completes the support.
Please consider that those kind of repositories are never closed
when calling close_all_repositories method and you have to call
remove_repository to effectively destroy all memory data.
Also, consider this feature EXPERIMENTAL and not for daily usage,
in other words, use it with brain ON!