[entropy.client.interfaces.dep] restrict tagged packages dependency match inside installed tags
This change affects dependency matching in a way that Entropy package tags
are used to reduce package matching scope when a tagged package is already
installed. For example, in Sabayon, when a new kernel is available, previous
Entropy versions used to pull it in. Avoiding to do this automatically and if
not required by user, makes possible to not risk to make a system unbootable,
especially when using proprietary, external video drivers.