[entropy.db] make EntropyRepository process-safe and thread-safe at the same time
This is kinda experimental but it seems to work just fine and, at the moment
there are no multithreaded accesses to EntropyRepository.
This patch makes write mutex in EntropyRepository useless because EntropyRepository
will automatically allocate new connections and cursors for every new
thread/pid coming in, avoiding to leak open files around.
SQLite3 is thread-safe when using one connection per thread, so... here it is!