[entropy.qa] add extra (ldd based) unresolved soname QA check
Using ldd to check if an executable has automagic dependencies
server-side, under controlled circumstances (such as checking against
the files provided by the package it belongs) is a good (not best)
way to make sure there are no breakages visible to user.
Unfortunately, these kind of missing dependencies have to be checked
and verified manually, besides, it's a good thing to have them
reported.