[equo] behaviour change: automatically bind non-tagged package strings to tagged
A simple use case: "equo install nvidia-drivers" will now pull in
packages providing the tag of the latest installed package if
available. In this way, newer kernel packages (for example) will
not be implicitly pulled in anymore. To override this: just append
a valid tag to your package string "nvidia-drivers#2.6.34-sabayon"