[entropy.client.interfaces.dep] add automatic conflicts detection, when package metadata doesn't list them.
It can happen that some installed packages conflict against newly
available packages, but the latter doesn't list the former as
conflict, resulting in unforeseen missing conflict information.
This is generally a QA issue in the package metadata, but it's better
to try to cope with it, even if a bit dangerous (effects of this
"auto-conflict" feature have to be tested in real world). For this
reason, one can override this feature by setting the env var:
ETP_DISABLE_AUTOCONFLICT=1 as well as turning off caching that could
compromise the results: ETP_NOCACHE=1