[entropy.client] quickpkg(): fix broken handling of hard-links
Since tar-1.24, tar supports hard links, same for python-2.6.6.
However, generate_package() was resetting the file type on the
tarinfo object, causing the extractor (either tar or python tarfile)
to handle the file like it was a 0-size one.
dev-vcs/git package was broken due to this and there are probably
others.