[entropy.client.interfaces.package] always run configuration files protection logic
When diffremoval is True (== package being updated), the configuration
files protection logic was skipped and every files removed regardless
user settings. This is wrong, since nowadays we store md5 of installed
files and we are able to know when a file is sitting on the fs unmodified,
thus decide to remove it automatically.
Also see bug #2354. Thanks Mitch Harder for reporting.