[entropy.client.interfaces.dep] filter or dependencies during reverse deps calculation
Let's say we have app-foo/one requiring python:2.6 or python:2.7.
We have both installed of them installed and we decide to drop
python:2.6. Without handling the case where either 2.6 or 2.7
are fine, the first occurence is pulled in, which is wrong.
This code changes the way retrieveReverseDependencies() work (
by including any or dependency in the reverse deps list) and
add the logic to handle the case in the reverse dependencies
calculation code.