[entropy.client.interfaces.dep] evaluate post-dependencies separately
Evaluating post-dependencies together with direct dependencies
could cause unexpected packages to be pulled in due to cycles in
the dependency graph. This commits makes possible to cut these
cycles by filtering out dependencies already pulled in, in the same
scope.

Example: nvidia-drivers pulls in nvidia-userspace which pulls in
nvidia-drivers as post-dependency. But given that we have more
nvidia-drivers packages in the same scope, this turned to be an issue