[entropy.client.interfaces.db] properly handle GPG signature install w/shadow repos
Shadow repositories are used to import other repos keys. So, if user
switch back to the original repo, Entropy is not able to import again
the same GPG key because gpg complains about having it already. So,
the entropy-side mapping (repo_id => key) cannot be updated easily.
Since it's impossible to reliably use gpg via fork()+exec(), it is
required to parse the current gpg key wanting to be imported, get its
fingerprint, look for repos with the same fingerprint and kill them from
the mapping. And in the end, retry the key installation.