[entropy.client.interfaces.dep] add support for ASAP scheduling of post-dependencies
This is a hack. As hack, it is hopefully going away soon.
This is a hack. It just hides broken dependencies (like libreoffice
vs jdom vs jdom-jaxen).
This is a hack. But Portage does ugly things.
This is a hack. But my hope is that it's going to be no longer required soon.