[entropy.server.mirrors] implement herustic able to use remote file copy when possible
When dealing with multiple repositories, it is possible that a package file
gets uploaded multiple times when moved from a repository to another.
In this scenario, there is a waste of upload bandwidth. This commit
implements a simple herustic that is able to determine if it's possible
to copy the file remotely instead of uploading it, if and only if:

- package file with the same name is already available on another
  tracked repository sharing the same mirror URI
- local package and remote package have the same md5 hash