[entropy.client.interfaces.methods] always use flock to handle lock file status
Make _check_pid_file_lock() use flock with LOCK_NB | LOCK_EX and LOCK_UN
to determine if the passed pid file is locked, and contains a different
pid value.
At the same time, unlock_resources() now removes the pid file before
releasing the lock associated with it. This makes possible to avoid
waiting 5 seconds when polling the same pid file to determine if it's
locked or not.