[entropy.client.interfaces.dep] introduce the "Slot intersection" feature
Certain dependency strings could have
cross-SLOT scope (multiple slots for same package are valid)
causing unwanted dependencies to be pulled in.
For example: if dependency is "dev-lang/python"
and we have dev-lang/python-2 installed, python-3
should be filtered out (if possible) by checking if
the installed best dependency match slot is still
available in repositories.
If it is, restrict the dependency scope to the intersection
between available SLOTs and installed SLOT.