[RigoDaemon] add deferred_shutdown() and shutdown() signals
Both can be used as replacement for the ping/pong protocol in order
to let RigoDaemon shutdown.
This is particularly useful for notification applets, in where holding
RigoDaemon instance is suboptimal.
On the other hand, letting RigoDaemon reload may then cause API
mismatches with these notification applets if they don't reload
as well.