[entropy.client] fix broken soname dependencies runtime routine wrt virtual pkgs
Virtual are not directly owning sonames reverse dependencies need.
For this reason, when a virtual package is encountered, its direct
dependencies must be taken into consideration in the soname bump
routine.
Moreover, pkgs requiring newer (or older?) sonames, must be bound
to soname providers on the graph, as direct dependency. Not doing
so would lead pkg_* phases to fail.

Scenario:
dev-libs/glib wants newer libffi.so.x, but the pkg depends against
virtual/libffi. If virtual/libffi dependencies are not taken into
consideration, _lookup_library_breakages() would fail to detect
the need of a bump.