[entropy.client] improve reliability of the dep conflict replacement routine
When an available conflict replacement is found, it is better to
enforce its installation before the actual conflicting pkg, in order
to avoid to have the conflict removed for some time during the
transaction phase. Moreover, if the transaction is interrupted
after the conflict removal but before the new pkg install, this
pkg won't be upgraded anymore, because it would become not installed.

Scenario:
net-tools conflicts with old sys-apps/openrc. sys-apps/openrc gets
removed during net-tools install phase. sys-apps/openrc is scheduled
for install a lot later. This also introduced a glitch with base-gcc,
triggering gcc-config, which complained about missing /etc/init.d/functions.sh