[matter] add support for "spm-repository-change-if-upstreamed"
If SPM repository changes are not allowed when a package needs to be bumped,
the package won't be bumped/recompiled.
However, there is a case in where the SPM repo change constraint loses its
importance, and it's when the current SPM repository no longer contains the
target package (thus, changes have been upstreamed or they're just not needed
anymore).
Using "spm-repository-change-if-upstreamed: yes" makes possible to ignore
SPM repo changes in this case.

The default behaviour does not change, spm-repository-change-if-upstreamed is
disabled by default (== no as default config value).