[matter] implement support for package dependencies grouping
Previously, each package in the particle was analyzed on its own and
spawning portage against multiple packages was not possible.
This commit introduces the ability to aggregate packages in particles
in "groups", so that they can be considered together in the same
portage merge transaction.

The previously exported environment variable MATTER_PACKAGE_NAME has
been renamed into MATTER_PACKAGE_NAMES and contains a space separated
list of package dependencies that are in the merge transaction.