[entropy.server] add drained and merged repositories dependencies testing
It happens that during our workflow, we bump packages into overlay repositories
like sabayon-limbo and sabayon-hell. However, potential dependencies issues
may pop up only when repositories are merged (their content drained) into
others. This features assumes that repositories are listed in order of priority,
where the first one is the base repo and others are layers on top of the first one.