[entropy.client.dep] __get_library_breakages: handle the new library dependencies case
It happened that __get_library_breakages only handled library bumps and
not also another important case: new library dependencies. This check is
very important at runtime, because it helps coping with unsatisfied
library dependencies (that might be satisfied at the package dependencies
level).
For instance, upgrading udev may fail because the new udevadm depends
against libblkid.so.1, which is a new dependency, while the util-linux
dependency metadatum is not enforcing the new version well enough.