[matter] improve ACCEPT_KEYWORDS setup basing on the "stable" parameter
This tweak makes possible to directly skip depgraph generation if
the root node (the dependency to check, last node of the toposorted
queue) is already up-to-date (wrt to stable: yes, for example).

This is what Portage does internally (using backupenv) and may break
in future releases without notice. So, keep the settings["ACCEPT_KEYWORDS"]
setup as a guard, to ensure the safety property.