[entropy.spm] PortagePlugin: drop the old license setup code, fix portage >=2.2.0_alpha189 support
The old license setup code is no longer needed since 2009. EAPI < 3
use ACCEPT_PROPERTIES="* -interactive", which is now set in the environment
(note that we may want to reduce the scope of the variable to portage.config
in the future). EAPI >= 3 use ACCEPT_LICENSE. check_license() in
eutils.eclass is gone as well.