[magneto] import entropy modules at upper level
Otherwise "entropy" module loading logic was delayed, excluding this
import (which was done early, before entropy_path_loader logic):

    from entropy.locks import SimpleFileLock

and it failed when system wide PYTHONPATH is dropped.

There was a comment not to do an early import due to --no-pid-handling.
The comment was added in:

    commit 59e5fd737e9360329d1a627ce49322f89274ca9d
    Date:   Tue Mar 1 13:13:18 2011 +0100

	[magneto] fix lock file handling

but then the logic was reworked in:

    commit 5d315c70bdbd455dc46f680b54275cb59853dd43
    Date:   Sun Feb 24 23:34:49 2013 +0100

	[entropy.const] drop /var/run/entropy/entropy.lock file and directory handling

and then passing --no-pid-handling (which was unused at this point) from
mangeto file was removed in:

    commit 83b8991c95d080973631a5c1e36af6a7573eeed9
    Date:   Wed Feb 27 20:59:14 2013 +0000

	[magneto] rewrite systray icons management, improve KDE, GNOME and MATE integration