[entropy.spm] adapt to qfile from portage-utils-0.80
In 0.74 -e was available, and in 0.80 it is not present.
In both versions, and with the same set of options used, -e and -v
provide the same result so -v is now used to work on both.

Changed in portage-utils, commit 951a8711a59b1a7d49125f5f5214ff1ae9e50074:

    qfile: drop non-functional --exact option

    Bug: https://bugs.gentoo.org/678632